requirements.txt
pip
discord.py
google
python-dotenv
psycopg2-binary==2.8.4
psycopg2==2.7.5